Restructure folders to use service poro facade design pattern
