+++ app/poros/movie_detail.rb
