+++ app/poros/notes.md
