+++ spec/rails_helper.rb
@@ -1,4 +1,3 @@
-require 'simplecov'
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
 ENV['RAILS_ENV'] ||= 'test'
